pyproject.toml
[project]
name = "stride-pi-bench"
version = "0.1.0"
description = "STRIDE-optimized Purple Agent for Pi-Bench (AgentBeats)"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2",
    "uvicorn[standard]>=0.20",
    "litellm>=1.0",
    "starlette>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
